\name{REGARMA-package}
\alias{REGARMA-package}
\alias{REGARMA}
\docType{package}
\title{Regularized estimation of a REGARMA model.}
\description{
  This packages provides functions to fit an autoregressive moving average model with
  time-dependent predictors (REGARMA) using a penalised likelihood approach under AIC, BIC 
  or GCV model selection criteria. 
}
\details{
  \tabular{ll}{
    Package: \tab REGARMA\cr
    Type: \tab Package\cr
    Version: \tab 0.2-5\cr
    Date: \tab 2014-01-15\cr
    License: \tab GPL2 or newer\cr
    LazyLoad: \tab yes\cr
  }
  
}
\author{
  Hamed Haselimashhadi <hamed.haselimashhadi@brunel.ac.uk>
  \cr Website: \url{www.hamedhaseli.webs.com}
}
\references{
  R. Tibshirani (1996), "Regression shrinkage and selection via the lasso", Journal of the Royal Statistical Society. Series B, 267-288.\cr
  H. Wang, G. Li, C. Tsai (2007), "Regression coefficient and autoregressive order shrinkage and selection via the lasso", Journal of the Royal Statistical Society. Series B, 69, 63-78.\cr
  H. Zou (2006), "The adaptive lasso and its oracle properties", Journal of the American statistical association, 476, 1418-1429.\cr
  Haselimashhadi, H. and Vinciotti, V. (2014) Penalised estimation of autoregressive moving average models with time-dependent predictors. Submitted.
}
\keyword{ package }
